feat: add AuthMiddleware

+++ src/Presentation/SocialNetwork.Presentation.Web/Controllers/UsersController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using SocialNetwork.Application.Contracts.Commands.Users;
 using SocialNetwork.Application.Contracts.Services;
+using SocialNetwork.Application.Services;
 using SocialNetwork.Presentation.Web.Contracts;
 
 namespace SocialNetwork.Presentation.Web.Controllers;
@@ -11,10 +12,12 @@ namespace SocialNetwork.Presentation.Web.Controllers;
 public class UsersController : ControllerBase
 {
     private readonly IUserService _userService;
+    private readonly CurrentUserManager _currentUserManager;
 
-    public UsersController(IUserService userService)
+    public UsersController(IUserService userService, CurrentUserManager userManager)
     {
         _userService = userService;
+        _currentUserManager = userManager;
     }
 
     [HttpGet]
@@ -60,9 +63,14 @@ public class UsersController : ControllerBase
     }
 
     [HttpPatch("{id:long}")]
-    [Authorize] // TODO: get current user
+    [Authorize]
     public async Task<ActionResult> ChangeUserName(long id, ChangeUserNameRequest request)
     {
+        if (_currentUserManager.CurrentUser?.Id != id)
+        {
+            return Forbid();
+        }
+        
         var response = await _userService.ChangeUserName(new(id, request.NewName));
 
         if (response is ChangeUserNameCommand.Response.InvalidRequest invalidRequest)
@@ -84,9 +92,14 @@ public class UsersController : ControllerBase
     }
 
     [HttpDelete("{id:long}")]
-    [Authorize] // TODO: get current user
+    [Authorize]
     public async Task<ActionResult> DeleteUser(long id)
     {
+        if (_currentUserManager.CurrentUser?.Id != id)
+        {
+            return Forbid();
+        }
+        
         var response = await _userService.DeleteUser(new(id));
 
         if (response is DeleteUserCommand.Response.NotFound)
